requirements.txt
apscheduler
google-api-python-client
pydantic
pygsheets
pyrogram
python-dotenv
sqlalchemy
tgcrypto
